Move followingStateToStyle to shared
