+++ shared/searchv3/result-row/index.js
@@ -4,21 +4,7 @@ import React from 'react'
 import {Box, Icon, ClickableBox, Text} from '../../common-adapters'
 import {globalColors, globalStyles, globalMargins} from '../../styles'
 import IconOrAvatar from '../icon-or-avatar'
-
-const followingStateToStyle = (followingState: Constants.FollowingState) => {
-  return {
-    Following: {
-      color: globalColors.green2,
-    },
-    NoState: {},
-    NotFollowing: {
-      color: globalColors.blue,
-    },
-    You: {
-      fontStyle: 'italic',
-    },
-  }[followingState]
-}
+import {followingStateToStyle} from '../shared'
 
 const Left = ({leftService, leftIcon, leftUsername, leftFollowingState}) => {
   return (
